Added Vault support
You can replace the <item stack> value in the trades.config with:
money: <double>

I've made it so that money acts like regular items for trading.

I've also added a customTradeMask field to the VillagerData that
makes it possible to work out which custom trade is 'covering' which
vanilla trade (as both the vanillaTrades and customTradeKeys are
ordered).

CustomTrades are now refreshed based on the key. That means that if you
make a change to a custom key in the config, it will be reflected
on all villagers that have that custom trade in the game.